+++ src/app/cdk/interstitials/interstitials.module.ts
@@ -13,15 +13,33 @@ import { MatLegacyButtonModule as MatButtonModule } from '@angular/material/lega
 import { InfoPanelModule } from '../info-panel/info-panel.module'
 import { A11yLinkModule } from '../a11y-link/a11y-link.module'
 import { ShareEmailsDomainsDialogComponent } from './share-emails-domains/share-emails-domains-dialog.component'
+import { AffiliationsInterstitialComponent } from './affiliations-interstitial/affiliations-interstitial.component'
+import { AffiliationsInterstitialDialogComponent } from './affiliations-interstitial/affiliations-interstitial-dialog.component'
+import { MatLegacyAutocompleteModule as MatAutocompleteModule } from '@angular/material/legacy-autocomplete'
+import { MatLegacyFormFieldModule } from '@angular/material/legacy-form-field'
+import { MatLegacySelectModule } from '@angular/material/legacy-select'
+import { SharedModule } from 'src/app/shared/shared.module'
+import { MAT_AUTOCOMPLETE_SCROLL_STRATEGY } from '@angular/material/autocomplete'
+import { MAT_SELECT_SCROLL_STRATEGY_PROVIDER } from '@angular/material/select'
+import { MatLegacyInputModule } from '@angular/material/legacy-input'
+import { MatLegacyProgressSpinnerModule } from '@angular/material/legacy-progress-spinner'
 
 @NgModule({
   declarations: [
     ShareEmailsDomainsComponent,
     ShareEmailsDomainsDialogComponent,
+    AffiliationsInterstitialComponent,
+    AffiliationsInterstitialDialogComponent,
   ],
   imports: [
     CommonModule,
+    SharedModule,
     MatLegacyCardModule,
+    MatAutocompleteModule,
+    MatLegacyFormFieldModule,
+    MatLegacySelectModule,
+    MatLegacyInputModule,
+    MatLegacyProgressSpinnerModule,
     MatIconModule,
     MatDividerModule,
     InfoDropDownModule,
@@ -33,6 +51,11 @@ import { ShareEmailsDomainsDialogComponent } from './share-emails-domains/share-
     InfoPanelModule,
     A11yLinkModule,
   ],
-  exports: [ShareEmailsDomainsComponent, ShareEmailsDomainsDialogComponent],
+  exports: [
+    ShareEmailsDomainsComponent,
+    ShareEmailsDomainsDialogComponent,
+    AffiliationsInterstitialComponent,
+    AffiliationsInterstitialDialogComponent,
+  ],
 })
 export class InterstitialsModule {}