+++ src/app/register/pages/register/register.component.ts
@@ -157,27 +157,7 @@ export class RegisterComponent implements OnInit, AfterViewInit {
         .subscribe((response) => {
           this.loading = false
           if (response.url) {
-            const analyticsReports: Observable<void>[] = []
-
-            analyticsReports.push(
-              this._googleTagManagerService.reportEvent(
-                'New-Registration',
-                this.requestInfoForm || 'Website'
-              )
-            )
-            forkJoin(analyticsReports)
-              .pipe(
-                catchError((err) =>
-                  this._errorHandler.handleError(
-                    err,
-                    ERROR_REPORT.STANDARD_NO_VERBOSE_NO_GA
-                  )
-                )
-              )
-              .subscribe(
-                () => this.afterRegisterRedirectionHandler(response),
-                () => this.afterRegisterRedirectionHandler(response)
-              )
+            this.afterRegisterRedirectionHandler(response)
           } else {
             this._errorHandler.handleError(
               new Error('registerUnexpectedConfirmation'),